Namespace for Spree 1.0
