--- app/models/gateway/stripe.rb
@@ -1,10 +1,0 @@
-class Gateway::Stripe < Gateway
-	preference :login, :string
-
-  def provider_class
-    ActiveMerchant::Billing::StripeGateway
-  end
-  def payment_profiles_supported?
-    false
-  end
-end
