+++ app/models/spree/gateway/stripe.rb
@@ -1,0 +1,12 @@
+module Spree
+  class Gateway::Stripe < Gateway
+  	preference :login, :string
+
+    def provider_class
+      ActiveMerchant::Billing::StripeGateway
+    end
+    def payment_profiles_supported?
+      false
+    end
+  end
+end
